+++ components/bio-formats/src/loci/formats/meta/AggregateMetadata.java
@@ -24,7 +24,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 /*-----------------------------------------------------------------------------
  *
  * THIS IS AUTOMATICALLY GENERATED CODE.  DO NOT MODIFY.
- * Created by curtis via MetadataAutogen on Oct 22, 2008 1:03:38 PM CDT
+ * Created by curtis via MetadataAutogen on Oct 17, 2008 1:25:59 AM CDT
  *
  *-----------------------------------------------------------------------------
  */
@@ -1384,12 +1384,12 @@ public class AggregateMetadata implements IMetadata {
   }
 
   /* @see MetadataRetrieve#getLogicalChannelPinholeSize(int, int) */
-  public Integer getLogicalChannelPinholeSize(int imageIndex, int logicalChannelIndex) {
+  public Float getLogicalChannelPinholeSize(int imageIndex, int logicalChannelIndex) {
     for (Iterator iter = delegates.iterator(); iter.hasNext();) {
       Object o = iter.next();
       if (o instanceof MetadataRetrieve) {
         MetadataRetrieve retrieve = (MetadataRetrieve) o;
-        Integer result = retrieve.getLogicalChannelPinholeSize(imageIndex, logicalChannelIndex);
+        Float result = retrieve.getLogicalChannelPinholeSize(imageIndex, logicalChannelIndex);
         if (result != null) return result;
       }
     }
@@ -3405,8 +3405,8 @@ public class AggregateMetadata implements IMetadata {
     }
   }
 
-  /* @see MetadataStore#setLogicalChannelPinholeSize(Integer, int, int) */
-  public void setLogicalChannelPinholeSize(Integer pinholeSize, int imageIndex, int logicalChannelIndex) {
+  /* @see MetadataStore#setLogicalChannelPinholeSize(Float, int, int) */
+  public void setLogicalChannelPinholeSize(Float pinholeSize, int imageIndex, int logicalChannelIndex) {
     for (Iterator iter = delegates.iterator(); iter.hasNext();) {
       Object o = iter.next();
       if (o instanceof MetadataStore) {
